src/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jackchat"
version = "0.1.0"
description = "A package for RAG-based chatbots using Azure AI Search and OpenAI."
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "langchain-openai",
    "langchain-community",
    "langchain-core",
    "tiktoken",
    "python-dotenv",
    "azure-search-documents"
]

[project.scripts]
rag-azure = "jackchat.cli:main"